+++ src/execution/createPipelineExecutor/20-executeTemplate.ts
@@ -1,5 +1,6 @@
 import { spaceTrim } from 'spacetrim';
 import type { Promisable, ReadonlyDeep, WritableDeep } from 'type-fest';
+import { MAX_EXECUTION_ATTEMPTS } from '../../config';
 import { extractParameterNamesFromTemplate } from '../../conversion/utils/extractParameterNamesFromTemplate';
 import { UnexpectedError } from '../../errors/UnexpectedError';
 import type { ExecutionReportJson } from '../../types/execution-report/ExecutionReportJson';
@@ -9,8 +10,7 @@ import type { TaskProgress } from '../../types/TaskProgress';
 import type { Parameters } from '../../types/typeAliases';
 import { difference } from '../../utils/sets/difference';
 import { union } from '../../utils/sets/union';
-import type { ExecutionTools } from '../ExecutionTools';
-import type { CreatePipelineExecutorSettings } from './00-CreatePipelineExecutorSettings';
+import { CreatePipelineExecutorOptions } from './00-CreatePipelineExecutorOptions';
 import { executeFormatSubvalues } from './30-executeFormatSubvalues';
 import { getReservedParametersForTemplate } from './getReservedParametersForTemplate';
 
@@ -19,7 +19,7 @@ import { getReservedParametersForTemplate } from './getReservedParametersForTemp
  *
  * @private internal type of `executeTemplate`
  */
-type executeSingleTemplateOptions = {
+type executeSingleTemplateOptions = CreatePipelineExecutorOptions & {
     /**
      * @@@
      */
@@ -38,17 +38,7 @@ type executeSingleTemplateOptions = {
     /**
      * @@@
      */
-    readonly tools: ExecutionTools;
-
-    /**
-     * @@@
-     */
     readonly onProgress: (taskProgress: TaskProgress) => Promisable<void>;
-
-    /**
-     * Settings for the pipeline executor
-     */
-    readonly settings: CreatePipelineExecutorSettings;
 
     /**
      * @@@
@@ -73,11 +63,10 @@ export async function executeTemplate(options: executeSingleTemplateOptions): Pr
         parametersToPass,
         tools,
         onProgress,
-        settings,
         $executionReport,
         pipelineIdentification,
+        maxExecutionAttempts = MAX_EXECUTION_ATTEMPTS,
     } = options;
-    const { maxExecutionAttempts } = settings;
 
     const name = `pipeline-executor-frame-${currentTemplate.name}`;
     const title = currentTemplate.title;
@@ -190,7 +179,6 @@ export async function executeTemplate(options: executeSingleTemplateOptions): Pr
         template: currentTemplate,
         preparedPipeline,
         tools,
-        settings,
         $executionReport,
         pipelineIdentification,
     });