+++ CHANGELOG.md
@@ -549,6 +549,7 @@ Knowledge scrapers [🐝]
 -   Rename `getLlmToolsForTestingAndScriptsAndPlayground` -> `$provideLlmToolsForTestingAndScriptsAndPlayground`
 -   Rename `getLlmToolsForCli` -> `$provideLlmToolsForCli`
 -   Change most `Array` -> `ReadonlyArray`
+-   Unite `CreatePipelineExecutorOptions` and `CreatePipelineExecutorSettings`
 
 ### `0.72.0` _(2024-09-)_